backend/requirements.txt
Django
djangorestframework 
djangorestframework-simplejwt 
dj-rest-auth 
django-allauth
requests
cryptography
rest_framework_simplejwt